requirements.txt
asgiref==3.3.1
astroid==2.5.6
boto3==1.19.8
botocore==1.22.8
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
clamd==1.0.2
colorama==0.4.4
coverage==5.5
crispy-forms-gds==0.2.2
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-nose==1.4.7
django-safe-filefield==0.3.1
django-storages==1.12.2
gunicorn==20.1.0
idna==2.10
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
nose==1.3.7
oauthlib==3.1.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.0.1
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
PyMySQL==1.0.2
python-dateutil==2.8.1
python-magic==0.4.22
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
django-heroku